Updated server to support more generic objects
